refactor(tentacle): merge pi-rpc-client into pi.ts (two sections, one file)
